Only store simple constants
closes #28

This is a bit of a "one fell swoop" fix

I'm still not sure why print tree was failing in the first place,
however I could nail it to "constants that reference motors that are not
in the axes or other constants"

I was also considering this change before I realized it solved this
particular issue because the constants produced from expressions are
often obtuse and make the autogenerated plots that much harder to parse.
(And are often supposed to be 0 since d1=d2 gets translated as d1-d2=0)

When I realized it would fix #28, I just decided to go with it.

The "simple" constant of "variable = #" is retained, as those are not
obtuse and can be genuinely helpful and have the same meaning between
scan constant and data constant.